backend/requirements.txt
# ══════════════════════════════════════════════════════════════
# PadiGuard AI — Unified Backend Dependencies
# Merged from diagnosis/requirements.txt + swarm/requirements.txt
# ══════════════════════════════════════════════════════════════

# ── Web Framework (Diagnosis) ─────────────────────────────────
fastapi>=0.115.0
uvicorn>=0.34.0
websockets>=13.0

# ── Google AI & ADK (Diagnosis) ───────────────────────────────
google-genai>=1.12.0
google-adk>=1.0.0
google-cloud-aiplatform>=1.80.0
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.18.0

# ── Genkit AI Framework (Swarm) ───────────────────────────────
genkit>=0.4.0,<0.5.0

# ── Firebase Admin SDK (Swarm) ────────────────────────────────
firebase-admin>=6.5.0

# ── Data Validation (Shared) ─────────────────────────────────
pydantic>=2.10.0
pydantic-settings>=2.7.0

# ── HTTP & MCP (Swarm) ───────────────────────────────────────
httpx>=0.27.0
mcp>=1.0.0

# ── Environment (Shared) ─────────────────────────────────────
python-dotenv>=1.0.0
certifi>=2024.2.2

# ── Testing ───────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
